Make this repo a working Python package (#1)
* DOC: Convert the documentation to numpydoc style

Add .gitignore

* BUG: Typo on the number of predictions

* DOC: Documentation about attributes, change array_like -> np.ndarray

* Change from using `np.dot` to use matrix multiplication operator

* Refactoring the exponential examples

* Apply Black.
* Apply RST.
* Rename the argument of the callback function to distinguish it from
the `geo` object used in the calculation.
* Define the dimension of the problem from the shape of the time
points and the initial parameter.

* Refactoring on the MMR example

* Apply Black
* Apply RST
* Rename the argument of the callback function to distinguish it from
the `geo` object used in the calculation.
* Define the dimension of the problem from the shape of the time
points and the initial parameter.

* Applay Black

* Initial commit to convert repo into a package

* Move Jacobian and Avv wrapper to finite_difference.py

I think this makes more sense since these derivative estimations use
finite difference method. Also, we are planning to add a wrapper to
JAX/autograd to calculation Jacobian and Avv.

* Add requirements in a separate file

And in setup.py, read the requirements from this file. Also, make
setup.py to read README.md to generate the descriptions.

* Change package name MBAM -> mbam

According to PEP8, package's name should be in lowercase.

* Update TODO list

* Add wrappers for JAX to calculate jacobian and Avv

And add the corresponding tests

* Skip AD tests if JAX is not installed

* Add random seed

* Update the tolerances to loosen the tests

* Rename model.py -> model_exp.py

* Add MIT License

* Update on README

* Add test for geodesic calculation

The test uses a linear model, since this is the simplest case.

* Remove M and N from Geodesic.__init__()

M is not used anywhere in the calculation. N can be retrieved from the
initial parameters.

* ENH: Added options to specify index and forward in initial_velocity

* Improve on how to compute Avv with jax

In this version, Avv is calculated by taking second derivative of F(x
+ v*s) with respect to s at s=0.0.

Co-authored-by: Yonatan Kurniawan <“kurniawanyo@outlook.com”>
Co-authored-by: yonatank <yonatank@byu.edu>