Fix the test for the initial velocity function
In this updated test, I just check it the velocity from
`initial_velocity` function is paralleel to the right singular vector
of the Jacobian. The direction might be the opposite to each other,
just because `initial_velocity` function returns the direction in
which the velocity increases, by checking the direction of the acceleration.