v0.1.0 - Defaults to integrating with constant parameter space velocity
* Modification to include tau (data space distance)

* Set the default `parameterspacenorm` to True.

With this setting, the plot of tau (data space arclength) vs
t (parameter space arclength) will plateaus if the geodesic encounters
a boundary.

* Remove installation with pip from README

We haven't uploaded this package to pypi. I will update this after we
register this package to pypi.

* Increase the version number to 0.1.0

* Return the `pip install mbam` in README

* Add a check if the velocity is zero.

This will speed up the calculation in some cases, and this practice is
also found in the Julia implementation.

* Update documentation

* Add a test for parameterspacenorm flag

* Fix the test for the initial velocity function

In this updated test, I just check it the velocity from
`initial_velocity` function is paralleel to the right singular vector
of the Jacobian. The direction might be the opposite to each other,
just because `initial_velocity` function returns the direction in
which the velocity increases, by checking the direction of the acceleration.

* Bump up scipy requirement to >=1.6.0

This version is required in the test to use
`scipy.integrate.cumulative_trapezoid` function.

Co-authored-by: yonatank <yonatank@byu.edu>